macos-port: LP64 type sweep + shadow/decal shim + WIP phase 1
Mechanical hardening of types that were ABI-fragile on LP64 (macOS arm64):
- DeathType/VeterancyLevel flag bit ops force 32-bit literal shifts mod-32
- RandomClass::Seed pinned to uint32_t (LCG state must round-trip 32 bits)
- WWLib/WWAudio/WW3D2/WWDownload struct fields holding 32-bit pixel
  colors, file offsets, datetimes, sound timestamps switched from
  `unsigned long`/`long` to fixed-width `uint32_t`/`int32_t`
- BorderColor.m_borderColor, render2d color array, ww3d stat counters,
  TARGA/BFIOFile/Buff/SysTimer/CRC fields all sized explicitly
- SHADOW_DECAL_VERTEX.diffuse pinned to uint32_t (D3DFVF_DIFFUSE expects
  4 bytes, LP64 inflated stride was making decals render as black)

Shadow-mapping shim work in dx8_stub/metal_backend (capture+replay
scaffolding, peter-pan fix, camera-forward + light-projection plumbing)
plus port-plan doc updates carry over from phase 1 sessions.

No behavioural change on Windows: every Apple-specific edit is guarded
by `#if defined(__APPLE__)`, the type swaps are size-equivalent (4 ->
4 bytes on both ABIs).

+++ Core/GameEngine/Source/GameClient/GUI/IMEManager.cpp
@@ -1349,7 +1349,8 @@ void IMEManager::updateCandidateList( Int candidateFlags  )
 		if ( candidateFlags & candidate )
 		{
 			Bool unicode = TRUE;
-			unsigned long listCount = 0;
+			// LP64 sweep: ImmGetCandidateListCount* takes LPDWORD = uint32_t* now.
+			DWORD listCount = 0;
 
 			Int size = ImmGetCandidateListCountW( m_context, &listCount );
