+++ Core/GameEngine/Source/GameNetwork/GameSpy/MainMenuUtils.cpp
@@ -733,7 +733,8 @@ DWORD WINAPI asyncGethostbynameThreadFunc( void * szName )
 int asyncGethostbyname(char * szName)
 {
 	static int            stat = 0;
-	static unsigned long  threadid;
+	// LP64 sweep: CreateThread takes LPDWORD = uint32_t* now.
+	static DWORD          threadid;
 
 	if( stat == 0 )
 	{
